refactor: move game layout and settings to JS
